Refactor into separate files and fix closure bug
