+++ Pod/Classes/Locksmith.swift
@@ -2,363 +2,60 @@ import UIKit
 
 public let LocksmithDefaultService = NSBundle.mainBundle().infoDictionary![String(kCFBundleIdentifierKey)] as? String ?? "com.locksmith.defaultService"
 
-// MARK: Accessible
-public enum LocksmithAccessibleOption: RawRepresentable {
-    case WhenUnlocked, AfterFirstUnlock, Always, WhenPasscodeSetThisDeviceOnly, WhenUnlockedThisDeviceOnly, AfterFirstUnlockThisDeviceOnly, AlwaysThisDeviceOnly
-    
-    public init?(rawValue: String) {
-        switch rawValue {
-        case String(kSecAttrAccessibleWhenUnlocked):
-            self = WhenUnlocked
-        case String(kSecAttrAccessibleAfterFirstUnlock):
-            self = AfterFirstUnlock
-        case String(kSecAttrAccessibleAlways):
-            self = Always
-        case String(kSecAttrAccessibleWhenPasscodeSetThisDeviceOnly):
-            self = WhenPasscodeSetThisDeviceOnly
-        case String(kSecAttrAccessibleWhenUnlockedThisDeviceOnly):
-            self = WhenUnlockedThisDeviceOnly
-        case String(kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly):
-            self = AfterFirstUnlockThisDeviceOnly
-        case String(kSecAttrAccessibleAlwaysThisDeviceOnly):
-            self = AlwaysThisDeviceOnly
-        default:
-            print("Accessible: invalid rawValue provided. Defaulting to Accessible.WhenUnlocked.")
-            self = WhenUnlocked
-        }
-    }
-    
-    public var rawValue: String {
-        switch self {
-        case .WhenUnlocked:
-            return String(kSecAttrAccessibleWhenUnlocked)
-        case .AfterFirstUnlock:
-            return String(kSecAttrAccessibleAfterFirstUnlock)
-        case .Always:
-            return String(kSecAttrAccessibleAlways)
-        case .WhenPasscodeSetThisDeviceOnly:
-            return String(kSecAttrAccessibleWhenPasscodeSetThisDeviceOnly)
-        case .WhenUnlockedThisDeviceOnly:
-            return String(kSecAttrAccessibleWhenUnlockedThisDeviceOnly)
-        case .AfterFirstUnlockThisDeviceOnly:
-            return String(kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly)
-        case .AlwaysThisDeviceOnly:
-            return String(kSecAttrAccessibleAlwaysThisDeviceOnly)
-        }
-    }
-}
-
-// MARK: Locksmith Error
-public enum LocksmithError: String, ErrorType {
-    case Allocate = "Failed to allocate memory."
-    case AuthFailed = "Authorization/Authentication failed."
-    case Decode = "Unable to decode the provided data."
-    case Duplicate = "The item already exists."
-    case InteractionNotAllowed = "Interaction with the Security Server is not allowed."
-    case NoError = "No error."
-    case NotAvailable = "No trust results are available."
-    case NotFound = "The item cannot be found."
-    case Param = "One or more parameters passed to the function were not valid."
-    case RequestNotSet = "The request was not set"
-    case TypeNotFound = "The type was not found"
-    case UnableToClear = "Unable to clear the keychain"
-    case Undefined = "An undefined error occurred"
-    case Unimplemented = "Function or operation not implemented."
-    
-    init?(fromStatusCode code: Int) {
-        switch code {
-        case Int(errSecAllocate):
-            self = Allocate
-        case Int(errSecAuthFailed):
-            self = AuthFailed
-        case Int(errSecDecode):
-            self = Decode
-        case Int(errSecDuplicateItem):
-            self = Duplicate
-        case Int(errSecInteractionNotAllowed):
-            self = InteractionNotAllowed
-        case Int(errSecItemNotFound):
-            self = NotFound
-        case Int(errSecNotAvailable):
-            self = NotAvailable
-        case Int(errSecParam):
-            self = Param
-        case Int(errSecUnimplemented):
-            self = Unimplemented
-        default:
-            return nil
-        }
-    }
-}
-
-// With thanks to http://iosdeveloperzone.com/2014/10/22/taming-foundation-constants-into-swift-enums/
-// MARK: Security Class
-public enum LocksmithSecurityClass: RawRepresentable {
-    case GenericPassword, InternetPassword, Certificate, Key, Identity
-    static let allClasses = [GenericPassword, InternetPassword, Certificate, Key, Identity]
-    
-    public init?(rawValue: String) {
-        switch rawValue {
-        case String(kSecClassGenericPassword):
-            self = GenericPassword
-        case String(kSecClassInternetPassword):
-            self = InternetPassword
-        case String(kSecClassCertificate):
-            self = Certificate
-        case String(kSecClassKey):
-            self = Key
-        case String(kSecClassIdentity):
-            self = Identity
-        default:
-            print("SecurityClass: Invalid raw value provided. Defaulting to .GenericPassword")
-            self = GenericPassword
-        }
-    }
-    
-    public var rawValue: String {
-        switch self {
-        case .GenericPassword:
-            return String(kSecClassGenericPassword)
-        case .InternetPassword:
-            return String(kSecClassInternetPassword)
-        case .Certificate:
-            return String(kSecClassCertificate)
-        case .Key:
-            return String(kSecClassKey)
-        case .Identity:
-            return String(kSecClassIdentity)
-        }
-    }
-}
-
-public enum LocksmithInternetAuthenticationType: RawRepresentable {
-    case NTLM, MSN, DPA, RPA, HTTPBasic, HTTPDigest, HTMLForm, Default
-    
-    public init?(rawValue: String) {
-        switch rawValue {
-        case String(kSecAttrAuthenticationTypeNTLM):
-            self = NTLM
-        case String(kSecAttrAuthenticationTypeMSN):
-            self = MSN
-        case String(kSecAttrAuthenticationTypeDPA):
-            self = DPA
-        case String(kSecAttrAuthenticationTypeRPA):
-            self = RPA
-        case String(kSecAttrAuthenticationTypeHTTPBasic):
-            self = HTTPBasic
-        case String(kSecAttrAuthenticationTypeHTTPDigest):
-            self = HTTPDigest
-        case String(kSecAttrAuthenticationTypeHTMLForm):
-            self = HTMLForm
-        case String(kSecAttrAuthenticationTypeDefault):
-            self = Default
-        default:
-            self = Default
-        }
-    }
-    
-    public var rawValue: String {
-        switch self {
-        case .NTLM:
-            return String(kSecAttrAuthenticationTypeNTLM)
-        case .MSN:
-            return String(kSecAttrAuthenticationTypeMSN)
-        case .DPA:
-            return String(kSecAttrAuthenticationTypeDPA)
-        case .RPA:
-            return String(kSecAttrAuthenticationTypeRPA)
-        case .HTTPBasic:
-            return String(kSecAttrAuthenticationTypeHTTPBasic)
-        case .HTTPDigest:
-            return String(kSecAttrAuthenticationTypeHTTPDigest)
-        case .HTMLForm:
-            return String(kSecAttrAuthenticationTypeHTMLForm)
-        case .Default:
-            return String(kSecAttrAuthenticationTypeDefault)
-        }
-    }
-    
-}
-
-public enum LocksmithInternetProtocol: RawRepresentable {
-    case FTP, FTPAccount, HTTP, IRC, NNTP, POP3, SMTP, SOCKS, IMAP, LDAP, AppleTalk, AFP, Telnet, SSH, FTPS, HTTPS, HTTPProxy, HTTPSProxy, FTPProxy, SMB, RTSP, RTSPProxy, DAAP, EPPC, IPP, NNTPS, LDAPS, TelnetS, IMAPS, IRCS, POP3S
-    
-    public init?(rawValue: String) {
-        switch rawValue {
-        case String(kSecAttrProtocolFTP):
-            self = FTP
-        case String(kSecAttrProtocolFTPAccount):
-            self = FTPAccount
-        case String(kSecAttrProtocolHTTP):
-            self = HTTP
-        case String(kSecAttrProtocolIRC):
-            self = IRC
-        case String(kSecAttrProtocolNNTP):
-            self = NNTP
-        case String(kSecAttrProtocolPOP3):
-            self = POP3
-        case String(kSecAttrProtocolSMTP):
-            self = SMTP
-        case String(kSecAttrProtocolSOCKS):
-            self = SOCKS
-        case String(kSecAttrProtocolIMAP):
-            self = IMAP
-        case String(kSecAttrProtocolLDAP):
-            self = LDAP
-        case String(kSecAttrProtocolAppleTalk):
-            self = AppleTalk
-        case String(kSecAttrProtocolAFP):
-            self = AFP
-        case String(kSecAttrProtocolTelnet):
-            self = Telnet
-        case String(kSecAttrProtocolSSH):
-            self = SSH
-        case String(kSecAttrProtocolFTPS):
-            self = FTPS
-        case String(kSecAttrProtocolHTTPS):
-            self = HTTPS
-        case String(kSecAttrProtocolHTTPProxy):
-            self = HTTPProxy
-        case String(kSecAttrProtocolHTTPSProxy):
-            self = HTTPSProxy
-        case String(kSecAttrProtocolFTPProxy):
-            self = FTPProxy
-        case String(kSecAttrProtocolSMB):
-            self = SMB
-        case String(kSecAttrProtocolRTSP):
-            self = RTSP
-        case String(kSecAttrProtocolRTSPProxy):
-            self = RTSPProxy
-        case String(kSecAttrProtocolDAAP):
-            self = DAAP
-        case String(kSecAttrProtocolEPPC):
-            self = EPPC
-        case String(kSecAttrProtocolIPP):
-            self = IPP
-        case String(kSecAttrProtocolNNTPS):
-            self = NNTPS
-        case String(kSecAttrProtocolLDAPS):
-            self = LDAPS
-        case String(kSecAttrProtocolTelnetS):
-            self = TelnetS
-        case String(kSecAttrProtocolIMAPS):
-            self = IMAPS
-        case String(kSecAttrProtocolIRCS):
-            self = IRCS
-        case String(kSecAttrProtocolPOP3S):
-            self = POP3S
-        default:
-            self = HTTP
-        }
-    }
-    
-    public var rawValue: String {
-        switch self {
-        case .FTP:
-            return String(kSecAttrProtocolFTP)
-        case .FTPAccount:
-            return String(kSecAttrProtocolFTPAccount)
-        case .HTTP:
-            return String(kSecAttrProtocolHTTP)
-        case .IRC:
-            return String(kSecAttrProtocolIRC)
-        case .NNTP:
-            return String(kSecAttrProtocolNNTP)
-        case .POP3:
-            return String(kSecAttrProtocolPOP3)
-        case .SMTP:
-            return String(kSecAttrProtocolSMTP)
-        case .SOCKS:
-            return String(kSecAttrProtocolSOCKS)
-        case .IMAP:
-            return String(kSecAttrProtocolIMAP)
-        case .LDAP:
-            return String(kSecAttrProtocolLDAP)
-        case .AppleTalk:
-            return String(kSecAttrProtocolAppleTalk)
-        case .AFP:
-            return String(kSecAttrProtocolAFP)
-        case .Telnet:
-            return String(kSecAttrProtocolTelnet)
-        case .SSH:
-            return String(kSecAttrProtocolSSH)
-        case .FTPS:
-            return String(kSecAttrProtocolFTPS)
-        case .HTTPS:
-            return String(kSecAttrProtocolHTTPS)
-        case .HTTPProxy:
-            return String(kSecAttrProtocolHTTPProxy)
-        case .HTTPSProxy:
-            return String(kSecAttrProtocolHTTPSProxy)
-        case .FTPProxy:
-            return String(kSecAttrProtocolFTPProxy)
-        case .SMB:
-            return String(kSecAttrProtocolSMB)
-        case .RTSP:
-            return String(kSecAttrProtocolRTSP)
-        case .RTSPProxy:
-            return String(kSecAttrProtocolRTSPProxy)
-        case .DAAP:
-            return String(kSecAttrProtocolDAAP)
-        case .EPPC:
-            return String(kSecAttrProtocolEPPC)
-        case .IPP:
-            return String(kSecAttrProtocolIPP)
-        case .NNTPS:
-            return String(kSecAttrProtocolNNTPS)
-        case .LDAPS:
-            return String(kSecAttrProtocolLDAPS)
-        case .TelnetS:
-            return String(kSecAttrProtocolTelnetS)
-        case .IMAPS:
-            return String(kSecAttrProtocolIMAPS)
-        case .IRCS:
-            return String(kSecAttrProtocolIRCS)
-        case .POP3S:
-            return String(kSecAttrProtocolPOP3S)
-        }
-    }
-}
-
 public typealias PerformRequestClosureType = (requestReference: CFDictionaryRef, inout result: AnyObject?) -> (OSStatus)
+
+
+// MARK: - Locksmith
+public struct Locksmith {
+    public static func loadDataForUserAccount(userAccount: String, inService service: String = LocksmithDefaultService) -> [String: AnyObject]? {
+        struct ReadRequest: GenericPasswordSecureStorable, ReadableSecureStorable {
+            let service: String
+            let account: String
+        }
+        
+        let request = ReadRequest(service: service, account: userAccount)
+        return request.readFromSecureStore()?.data
+    }
+    
+    public static func saveData(data: [String: AnyObject], forUserAccount userAccount: String, inService service: String = LocksmithDefaultService) throws {
+        struct CreateRequest: GenericPasswordSecureStorable, CreateableSecureStorable {
+            let service: String
+            let account: String
+            let data: [String: AnyObject]
+        }
+        
+        let request = CreateRequest(service: service, account: userAccount, data: data)
+        return try request.createInSecureStore()
+    }
+    
+    public static func deleteDataForUserAccount(userAccount: String, inService service: String = LocksmithDefaultService) throws {
+        struct DeleteRequest: GenericPasswordSecureStorable, DeleteableSecureStorable {
+            let service: String
+            let account: String
+        }
+        
+        let request = DeleteRequest(service: service, account: userAccount)
+        return try request.deleteFromSecureStore()
+    }
+    
+    public static func updateData(data: [String: AnyObject], forUserAccount userAccount: String, inService service: String = LocksmithDefaultService) throws {
+        // Delete and then re-save
+        do {
+            try Locksmith.deleteDataForUserAccount(userAccount, inService: service)
+        } catch {
+            // Deletion is likely to fail if the piece of data doesn't exist yet.
+            // This doesn't matter--we only tell the user about errors on the save request.
+        }
+        
+        return try Locksmith.saveData(data, forUserAccount: userAccount, inService: service)
+    }
+}
 
 // MARK: - SecureStorable
 /// The base protocol that indicates conforming types will have the ability to be stored in a secure storage container, such as the iOS keychain.
 public protocol SecureStorable {
-//    var performRequestClosure: PerformRequestClosureType { get }
-//    var asSecureStoragePropertyDictionary: [String: AnyObject] { get }
     var accessible: LocksmithAccessibleOption? { get }
     var accessGroup: String? { get }
-}
-
-public protocol SecureStorableResultType: SecureStorable {
-    var resultDictionary: [String: AnyObject] { get }
-    var data: [String: AnyObject]? { get }
-}
-
-public extension SecureStorableResultType {
-    var performRequestClosure: PerformRequestClosureType {
-        // TODO: This is pretty bad
-        return { (_, _) in
-            return errSecSuccess
-        }
-    }
-    
-    var asSecureStoragePropertyDictionary: [String: AnyObject] {
-        return [String: AnyObject]()
-    }
-    
-    var resultDictionary: [String: AnyObject] {
-        return [String: AnyObject]()
-    }
-    
-    var data: [String: AnyObject]? {
-        guard let aData = resultDictionary[String(kSecValueData)] as? NSData else {
-            return nil
-        }
-        
-        return NSKeyedUnarchiver.unarchiveObjectWithData(aData) as? [String: AnyObject]
-    }
 }
 
 public extension SecureStorable {
@@ -396,11 +93,42 @@ public extension SecureStorable {
             return nil
         }
         
-        guard let data = dictionary[String(kSecValueData)] as? NSData else {
+        if dictionary[String(kSecValueData)] as? NSData == nil {
             return nil
         }
         
         return result as? [String: AnyObject]
+    }
+}
+
+public extension SecureStorable where Self : InternetPasswordSecureStorable {
+    private var internetPasswordBaseStoragePropertyDictionary: [String: AnyObject] {
+        var dictionary = [String: AnyObject]()
+        
+        // add in whatever turns out to be required...
+        dictionary[String(kSecAttrServer)] = self.server
+        dictionary[String(kSecAttrPort)] = self.port
+        dictionary[String(kSecAttrProtocol)] = self.internetProtocol.rawValue
+        dictionary[String(kSecAttrAuthenticationType)] = self.authenticationType.rawValue
+        dictionary[String(kSecAttrSecurityDomain)] = self.securityDomain
+        dictionary[String(kSecAttrPath)] = self.path
+        dictionary[String(kSecClass)] = LocksmithSecurityClass.InternetPassword.rawValue
+        
+        let toMergeWith = [
+            self.accountSecureStoragePropertyDictionary,
+            self.describableSecureStoragePropertyDictionary,
+            self.commentableSecureStoragePropertyDictionary,
+            self.creatorDesignatableSecureStoragePropertyDictionary,
+            self.typeDesignatableSecureStoragePropertyDictionary,
+            self.isInvisibleSecureStoragePropertyDictionary,
+            self.isNegativeSecureStoragePropertyDictionary
+        ]
+        
+        for dict in toMergeWith {
+            dictionary = Dictionary(initial: dictionary, toMerge: dict)
+        }
+        
+        return dictionary
     }
 }
 
@@ -603,6 +331,36 @@ public extension GenericPasswordSecureStorableResultType {
     }
 }
 
+public extension SecureStorable where Self : GenericPasswordSecureStorable {
+    private var genericPasswordBaseStoragePropertyDictionary: [String: AnyObject] {
+        var dictionary = [String: AnyObject?]()
+        
+        dictionary[String(kSecAttrService)] = self.service
+        dictionary[String(kSecAttrGeneric)] = self.generic
+        dictionary[String(kSecClass)] = LocksmithSecurityClass.GenericPassword.rawValue
+        
+        dictionary = Dictionary(initial: dictionary, toMerge: self.describableSecureStoragePropertyDictionary)
+        
+        let toMergeWith = [
+            self.secureStorableBaseStoragePropertyDictionary,
+            self.accountSecureStoragePropertyDictionary,
+            self.describableSecureStoragePropertyDictionary,
+            self.commentableSecureStoragePropertyDictionary,
+            self.creatorDesignatableSecureStoragePropertyDictionary,
+            self.typeDesignatableSecureStoragePropertyDictionary,
+            self.labellableSecureStoragePropertyDictionary,
+            self.isInvisibleSecureStoragePropertyDictionary,
+            self.isNegativeSecureStoragePropertyDictionary
+        ]
+        
+        for dict in toMergeWith {
+            dictionary = Dictionary(initial: dictionary, toMerge: dict)
+        }
+        
+        return Dictionary(withoutOptionalValues: dictionary)
+    }
+}
+
 // MARK: - InternetPasswordSecureStorable
 /// A protocol that indicates a type conforms to the requirements of an internet password in a secure storage container.
 public protocol InternetPasswordSecureStorable: AccountBasedSecureStorable, DescribableSecureStorable, CommentableSecureStorable, CreatorDesignatableSecureStorable, TypeDesignatableSecureStorable, IsInvisibleAssignableSecureStorable, IsNegativeAssignableSecureStorable {
@@ -664,89 +422,27 @@ public protocol KeySecureStorable: SecureStorable {}
 /// Conformance to this protocol indicates that your type is able to be created and saved to a secure storage container.
 public protocol CreateableSecureStorable: SecureStorable {
     var data: [String: AnyObject] { get }
+    var performCreateRequestClosure: PerformRequestClosureType { get }
     func createInSecureStore() throws
 }
 
 // MARK: - ReadableSecureStorable
 /// Conformance to this protocol indicates that your type is able to be read from a secure storage container.
 public protocol ReadableSecureStorable: SecureStorable {
+    var performReadRequestClosure: PerformRequestClosureType { get }
     func readFromSecureStore() -> SecureStorableResultType?
 }
 
-
-// MARK: - DeleteableSecureStorable
-/// Conformance to this protocol indicates that your type is able to be deleted from a secure storage container.
-public protocol DeleteableSecureStorable: SecureStorable {
-    func deleteFromSecureStore() throws
-}
-
-public extension Dictionary {
-    init(withoutOptionalValues initial: Dictionary<Key, Value?>) {
-        self = [Key: Value]()
-        for pair in initial {
-            if pair.1 != nil {
-                self[pair.0] = pair.1!
-            }
+public extension ReadableSecureStorable {
+    var performReadRequestClosure: PerformRequestClosureType {
+        return { (requestReference: CFDictionaryRef, inout result: AnyObject?) in
+            return withUnsafeMutablePointer(&result) { SecItemCopyMatching(requestReference, UnsafeMutablePointer($0)) }
         }
     }
     
-    init(pairs: [(Key, Value)]) {
-        self = [Key: Value]()
-        pairs.forEach { (k, v) -> () in
-            self[k] = v
-        }
-    }
-    
-    init(initial: Dictionary<Key, Value>, toMerge: Dictionary<Key, Value>) {
-        self = Dictionary<Key, Value>()
-        
-        for pair in initial {
-            self[pair.0] = pair.1
-        }
-        
-        for pair in toMerge {
-            self[pair.0] = pair.1
-        }
-    }
-}
-
-// MARK: - Default property dictionaries
-
-public extension SecureStorable where Self : GenericPasswordSecureStorable {
-    private var genericPasswordBaseStoragePropertyDictionary: [String: AnyObject] {
-        var dictionary = [String: AnyObject?]()
-        
-        dictionary[String(kSecAttrService)] = self.service
-        dictionary[String(kSecAttrGeneric)] = self.generic
-        dictionary[String(kSecClass)] = LocksmithSecurityClass.GenericPassword.rawValue
-        
-        dictionary = Dictionary(initial: dictionary, toMerge: self.describableSecureStoragePropertyDictionary)
-        
-        let toMergeWith = [
-            self.secureStorableBaseStoragePropertyDictionary,
-            self.accountSecureStoragePropertyDictionary,
-            self.describableSecureStoragePropertyDictionary,
-            self.commentableSecureStoragePropertyDictionary,
-            self.creatorDesignatableSecureStoragePropertyDictionary,
-            self.typeDesignatableSecureStoragePropertyDictionary,
-            self.labellableSecureStoragePropertyDictionary,
-            self.isInvisibleSecureStoragePropertyDictionary,
-            self.isNegativeSecureStoragePropertyDictionary
-        ]
-        
-        for dict in toMergeWith {
-            dictionary = Dictionary(initial: dictionary, toMerge: dict)
-        }
-        
-        return Dictionary(withoutOptionalValues: dictionary)
-    }
-}
-
-public extension CreateableSecureStorable where Self : GenericPasswordSecureStorable {
-    var asCreateableSecureStoragePropertyDictionary: [String: AnyObject] {
-        var old = self.genericPasswordBaseStoragePropertyDictionary
-        old[String(kSecValueData)] = NSKeyedArchiver.archivedDataWithRootObject(self.data)
-        return old
+    func readFromSecureStore() -> SecureStorableResultType? {
+        // This must be implemented here so that we can properly override it in the type-specific implementations
+        return nil
     }
 }
 
@@ -761,54 +457,6 @@ public extension ReadableSecureStorable where Self : GenericPasswordSecureStorab
     }
 }
 
-public extension DeleteableSecureStorable where Self : GenericPasswordSecureStorable {
-    var asDeleteableSecureStoragePropertyDictionary: [String: AnyObject] {
-        return self.genericPasswordBaseStoragePropertyDictionary
-    }
-}
-
-
-// MARK: InternetPasswordSecureStorable
-
-public extension SecureStorable where Self : InternetPasswordSecureStorable {
-    private var internetPasswordBaseStoragePropertyDictionary: [String: AnyObject] {
-        var dictionary = [String: AnyObject]()
-        
-        // add in whatever turns out to be required...
-        dictionary[String(kSecAttrServer)] = self.server
-        dictionary[String(kSecAttrPort)] = self.port
-        dictionary[String(kSecAttrProtocol)] = self.internetProtocol.rawValue
-        dictionary[String(kSecAttrAuthenticationType)] = self.authenticationType.rawValue
-        dictionary[String(kSecAttrSecurityDomain)] = self.securityDomain
-        dictionary[String(kSecAttrPath)] = self.path
-        dictionary[String(kSecClass)] = LocksmithSecurityClass.InternetPassword.rawValue
-        
-        let toMergeWith = [
-            self.accountSecureStoragePropertyDictionary,
-            self.describableSecureStoragePropertyDictionary,
-            self.commentableSecureStoragePropertyDictionary,
-            self.creatorDesignatableSecureStoragePropertyDictionary,
-            self.typeDesignatableSecureStoragePropertyDictionary,
-            self.isInvisibleSecureStoragePropertyDictionary,
-            self.isNegativeSecureStoragePropertyDictionary
-        ]
-        
-        for dict in toMergeWith {
-            dictionary = Dictionary(initial: dictionary, toMerge: dict)
-        }
-        
-        return dictionary
-    }
-}
-
-public extension CreateableSecureStorable where Self : InternetPasswordSecureStorable {
-    var asCreateableSecureStoragePropertyDictionary: [String: AnyObject] {
-        var old = self.internetPasswordBaseStoragePropertyDictionary
-        old[String(kSecValueData)] = NSKeyedArchiver.archivedDataWithRootObject(self.data)
-        return old
-    }
-}
-
 public extension ReadableSecureStorable where Self : InternetPasswordSecureStorable {
     var asReadableSecureStoragePropertyDictionary: [String: AnyObject] {
         var old = self.internetPasswordBaseStoragePropertyDictionary
@@ -816,49 +464,6 @@ public extension ReadableSecureStorable where Self : InternetPasswordSecureStora
         old[String(kSecMatchLimit)] = kSecMatchLimitOne
         old[String(kSecReturnAttributes)] = kCFBooleanTrue
         return old
-    }
-}
-
-public extension DeleteableSecureStorable where Self : InternetPasswordSecureStorable {
-    var asDeleteableSecureStoragePropertyDictionary: [String: AnyObject] {
-        return self.internetPasswordBaseStoragePropertyDictionary
-    }
-}
-
-public extension CreateableSecureStorable {
-    var performCreateRequestClosure: PerformRequestClosureType {
-        return { (requestReference: CFDictionaryRef, inout result: AnyObject?) in
-            return withUnsafeMutablePointer(&result) { SecItemAdd(requestReference, UnsafeMutablePointer($0)) }
-        }
-    }
-    
-    func createInSecureStore() throws {
-//        try performSecureStorageAction(performCreateRequestClosure, secureStoragePropertyDictionary: self.asC)
-    }
-}
-
-public extension CreateableSecureStorable where Self : InternetPasswordSecureStorable {
-    func createInSecureStore() throws {
-        try performSecureStorageAction(performCreateRequestClosure, secureStoragePropertyDictionary: self.asCreateableSecureStoragePropertyDictionary)
-    }
-}
-
-public extension CreateableSecureStorable where Self : GenericPasswordSecureStorable {
-    func createInSecureStore() throws {
-        try performSecureStorageAction(performCreateRequestClosure, secureStoragePropertyDictionary: self.asCreateableSecureStoragePropertyDictionary)
-    }
-}
-
-public extension ReadableSecureStorable {
-    var performReadRequestClosure: PerformRequestClosureType {
-        return { (requestReference: CFDictionaryRef, inout result: AnyObject?) in
-            return withUnsafeMutablePointer(&result) { SecItemCopyMatching(requestReference, UnsafeMutablePointer($0)) }
-        }
-    }
-    
-    func readFromSecureStore() -> SecureStorableResultType? {
-        // This must be implemented here so that we can properly override it in the type-specific implementations
-        return nil
     }
 }
 
@@ -878,10 +483,6 @@ public extension ReadableSecureStorable where Self : GenericPasswordSecureStorab
     }
 }
 
-struct InternetPasswordResult: InternetPasswordSecureStorableResultType {
-    var resultDictionary: [String: AnyObject]
-}
-
 public extension ReadableSecureStorable where Self : InternetPasswordSecureStorable {
     func readFromSecureStore() -> InternetPasswordSecureStorableResultType? {
         do {
@@ -894,11 +495,69 @@ public extension ReadableSecureStorable where Self : InternetPasswordSecureStora
     }
 }
 
+
+// MARK: - DeleteableSecureStorable
+/// Conformance to this protocol indicates that your type is able to be deleted from a secure storage container.
+public protocol DeleteableSecureStorable: SecureStorable {
+    var performDeleteRequestClosure: PerformRequestClosureType { get }
+    func deleteFromSecureStore() throws
+}
+
+// MARK: - Default property dictionaries
+
+public extension CreateableSecureStorable where Self : GenericPasswordSecureStorable {
+    var asCreateableSecureStoragePropertyDictionary: [String: AnyObject] {
+        var old = self.genericPasswordBaseStoragePropertyDictionary
+        old[String(kSecValueData)] = NSKeyedArchiver.archivedDataWithRootObject(self.data)
+        return old
+    }
+}
+
+public extension CreateableSecureStorable where Self : GenericPasswordSecureStorable {
+    func createInSecureStore() throws {
+        try performSecureStorageAction(performCreateRequestClosure, secureStoragePropertyDictionary: self.asCreateableSecureStoragePropertyDictionary)
+    }
+}
+
+public extension CreateableSecureStorable where Self : InternetPasswordSecureStorable {
+    var asCreateableSecureStoragePropertyDictionary: [String: AnyObject] {
+        var old = self.internetPasswordBaseStoragePropertyDictionary
+        old[String(kSecValueData)] = NSKeyedArchiver.archivedDataWithRootObject(self.data)
+        return old
+    }
+}
+
+public extension CreateableSecureStorable {
+    var performCreateRequestClosure: PerformRequestClosureType {
+        return { (requestReference: CFDictionaryRef, inout result: AnyObject?) in
+            return withUnsafeMutablePointer(&result) { SecItemAdd(requestReference, UnsafeMutablePointer($0)) }
+        }
+    }
+}
+
+public extension CreateableSecureStorable where Self : InternetPasswordSecureStorable {
+    func createInSecureStore() throws {
+        try performSecureStorageAction(performCreateRequestClosure, secureStoragePropertyDictionary: self.asCreateableSecureStoragePropertyDictionary)
+    }
+}
+
 public extension DeleteableSecureStorable {
     var performDeleteRequestClosure: PerformRequestClosureType {
         return { (requestReference, _) in
             return SecItemDelete(requestReference)
         }
+    }
+}
+
+public extension DeleteableSecureStorable where Self : GenericPasswordSecureStorable {
+    var asDeleteableSecureStoragePropertyDictionary: [String: AnyObject] {
+        return self.genericPasswordBaseStoragePropertyDictionary
+    }
+}
+
+public extension DeleteableSecureStorable where Self : InternetPasswordSecureStorable {
+    var asDeleteableSecureStoragePropertyDictionary: [String: AnyObject] {
+        return self.internetPasswordBaseStoragePropertyDictionary
     }
 }
 
@@ -914,48 +573,26 @@ public extension DeleteableSecureStorable where Self : InternetPasswordSecureSto
     }
 }
 
-// MARK: - Locksmith
-public struct Locksmith {
-    public static func loadDataForUserAccount(userAccount: String, inService service: String = LocksmithDefaultService) -> [String: AnyObject]? {
-        struct ReadRequest: GenericPasswordSecureStorable, ReadableSecureStorable {
-            let service: String
-            let account: String
-        }
-        
-        let request = ReadRequest(service: service, account: userAccount)
-        return request.readFromSecureStore()?.data
+// MARK: ResultTypes
+public protocol SecureStorableResultType: SecureStorable {
+    var resultDictionary: [String: AnyObject] { get }
+    var data: [String: AnyObject]? { get }
+}
+
+struct InternetPasswordResult: InternetPasswordSecureStorableResultType {
+    var resultDictionary: [String: AnyObject]
+}
+
+public extension SecureStorableResultType {
+    var resultDictionary: [String: AnyObject] {
+        return [String: AnyObject]()
     }
     
-    public static func saveData(data: [String: AnyObject], forUserAccount userAccount: String, inService service: String = LocksmithDefaultService) throws {
-        struct CreateRequest: GenericPasswordSecureStorable, CreateableSecureStorable {
-            let service: String
-            let account: String
-            let data: [String: AnyObject]
+    var data: [String: AnyObject]? {
+        guard let aData = resultDictionary[String(kSecValueData)] as? NSData else {
+            return nil
         }
         
-        let request = CreateRequest(service: service, account: userAccount, data: data)
-        return try request.createInSecureStore()
-    }
-    
-    public static func deleteDataForUserAccount(userAccount: String, inService service: String = LocksmithDefaultService) throws {
-        struct DeleteRequest: GenericPasswordSecureStorable, DeleteableSecureStorable {
-            let service: String
-            let account: String
-        }
-        
-        let request = DeleteRequest(service: service, account: userAccount)
-        return try request.deleteFromSecureStore()
-    }
-    
-    public static func updateData(data: [String: AnyObject], forUserAccount userAccount: String, inService service: String = LocksmithDefaultService) throws {
-        // Delete and then re-save
-        do {
-            try Locksmith.deleteDataForUserAccount(userAccount, inService: service)
-        } catch {
-            // Deletion is likely to fail if the piece of data doesn't exist yet.
-            // This doesn't matter--we only tell the user about errors on the save request.
-        }
-        
-        return try Locksmith.saveData(data, forUserAccount: userAccount, inService: service)
+        return NSKeyedUnarchiver.unarchiveObjectWithData(aData) as? [String: AnyObject]
     }
 }
